change everything ot int

+++ src/main/java/addToCart/impl/Item.java
@@ -2,13 +2,13 @@ package addToCart.impl;
 
 public class Item {
 
-    private  long id;
+    private  int id;
     private  String name;
     private  double price;
     private  String description;
     //menu_id
 
-    public Item(long id, String name, double price, String description) {
+    public Item(int id, String name, double price, String description) {
         this.id = id;
         this.name = name;
         this.price = price;
